+++ mods/Farming/mod/prototypes/recipe/root_recipe.lua
@@ -3,7 +3,7 @@ local patch_data = require("data.patch_data")
 
 for _, patch in pairs(patch_data) do
   create_recipe({
-    name = "rs-"..patch.name.."-root",
+    name = "rs-"..patch.name.."-roots",
     category = get_vanilla_name("rs-"..patch.name.."-farming"),
     energy_required = patch.mining_time,
     ingredients = {