+++ mods/Farming/CHANGELOG.md
@@ -3,6 +3,9 @@
 - Added oak, willow, maple, yew and magic seeds, roots and patches
 - Within the patches you can grow logs (Requires RsWoodcutting) and roots
 - Added compost and compost bins. Some items can now be turned into compost, those items act as "compost fuel". Compost and supercompost increase yield on patches.
+- Added technologies for tree growing (very nice tree of trees!)
+- Added recipes to grow trees with compost
+
 
 ## [0.1.0] - 2019-07-26 - LeonSkills
 ### Added